move loss helpers into pca module
